requirements.txt
Flask==3.1.3
gunicorn==23.0.0
playwright==1.58.0
python-dotenv==1.0.1
